+++ packages/inline-choice/configure/src/root.jsx
@@ -1,57 +1,65 @@
 import React from 'react';
 import Main from './main';
 import cloneDeep from 'lodash/cloneDeep';
+import PropTypes from 'prop-types';
+import { choiceUtils as utils } from '@pie-lib/config-ui';
 
 export default class Root extends React.Component {
+  static propTypes = {
+    model: PropTypes.object.isRequired,
+    onModelChanged: PropTypes.func.isRequired
+  };
   constructor(props) {
     super(props);
     this.state = {
       model: props.model
-    }
+    };
   }
 
   handleModelChange() {
-    this.props.onModelChanged(this.state.model);
+    const { onModelChanged } = this.props;
+    onModelChanged(this.state.model);
   }
 
   update(model) {
     this.setState({ model }, () => {
       this.handleModelChange();
-    })
+    });
   }
 
   onAddChoice = () => {
-    const update = cloneDeep(this.state.model);
+    const { model } = this.state;
+    const update = cloneDeep(model);
     update.choices.push({
       correct: false,
-      value: '',
+      value: utils.firstAvailableIndex(model.choices.map(c => c.value), 0),
       feedback: { type: 'default' },
       label: ''
     });
 
     this.update(update);
-  }
+  };
 
   onChoiceChange = (index, newChoice) => {
     const update = cloneDeep(this.state.model);
     if (newChoice.correct) {
-      update.choices.forEach(c => c.correct = false);
+      update.choices.forEach(c => (c.correct = false));
     }
     update.choices.splice(index, 1, newChoice);
     this.update(update);
-  }
+  };
 
-  onRemoveChoice = (indexToRemove) => {
+  onRemoveChoice = indexToRemove => {
     let update = cloneDeep(this.state.model);
     update.choices.splice(indexToRemove, 1);
     this.update(update);
-  }
+  };
 
-  onPromptChange = (prompt) => {
+  onPromptChange = prompt => {
     let update = cloneDeep(this.state.model);
     update.prompt = prompt;
     this.update(update);
-  }
+  };
 
   render() {
     return (