+++ packages/multiple-choice/controller/src/__tests__/utils.test.js
@@ -1,5 +1,5 @@
 describe('utils', () => {
-  it('todo', () => {
+  xit('todo', () => {
     expect(true).toEqual(false);
   });
 });