+++ packages/placement-ordering/configure/src/index.js
@@ -1,32 +1,32 @@
-import { DeleteImageEvent, InsertImageEvent, ModelUpdatedEvent } from '@pie-framework/pie-configure-events';
+import {
+  DeleteImageEvent,
+  InsertImageEvent,
+  ModelUpdatedEvent
+} from '@pie-framework/pie-configure-events';
 
 import Main from './main';
 import React from 'react';
 import ReactDOM from 'react-dom';
-import debug from 'debug';
-
-const log = debug('pie-element:placement-ordering');
 
 export default class PlacementOrdering extends HTMLElement {
-
   constructor() {
     super();
     this.onModelChange = (model, resetSession) => {
       this._model = model;
       this.dispatchUpdate(resetSession);
-    }
+    };
 
-    this.insertImage = (handler) => {
+    this.insertImage = handler => {
       this.dispatchEvent(new InsertImageEvent(handler));
-    }
+    };
 
     this.deleteImage = (src, done) => {
       this.dispatchEvent(new DeleteImageEvent(src, done));
-    }
+    };
   }
 
   dispatchUpdate(reset) {
-    const detail = { update: this._model, reset }
+    const detail = { update: this._model, reset };
     this.dispatchEvent(new ModelUpdatedEvent(this._model, reset));
   }