+++ CHANGELOG.md
@@ -3,6 +3,76 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="1.5.0"></a>
+# [1.5.0](https://github.com/pie-framework/pie-elements/compare/v1.4.1...v1.5.0) (2018-04-17)
+
+
+### Bug Fixes
+
+* **dependencies:** update dependencies ([c774453](https://github.com/pie-framework/pie-elements/commit/c774453))
+* **icon:** use [@material-ui](https://github.com/material-ui)/icons package ([9e45c1a](https://github.com/pie-framework/pie-elements/commit/9e45c1a))
+
+
+### Features
+
+* **dependencie:** use latest [@pie-ui](https://github.com/pie-ui)/text-entry ([d6ef0cb](https://github.com/pie-framework/pie-elements/commit/d6ef0cb))
+* **dependencies:** use [@pie-ui](https://github.com/pie-ui)/calculator@2.x ([6011077](https://github.com/pie-framework/pie-elements/commit/6011077))
+* **dependencies:** use latest [@pie-ui](https://github.com/pie-ui)/multiple-choice ([9b73dbb](https://github.com/pie-framework/pie-elements/commit/9b73dbb))
+* **dependencies:** use latest [@pie-ui](https://github.com/pie-ui)/number-line ([e8fcdba](https://github.com/pie-framework/pie-elements/commit/e8fcdba))
+* **dependencies:** use latest [@pie-ui](https://github.com/pie-ui)/placement-ordering ([fd9d12d](https://github.com/pie-framework/pie-elements/commit/fd9d12d))
+
+
+
+
+<a name="1.4.1"></a>
+## [1.4.1](https://github.com/pie-framework/pie-elements/compare/v1.4.0...v1.4.1) (2018-04-17)
+
+
+
+
+**Note:** Version bump only for package undefined
+
+<a name="1.4.0"></a>
+# [1.4.0](https://github.com/pie-framework/pie-elements/compare/v1.3.0...v1.4.0) (2018-04-17)
+
+
+### Bug Fixes
+
+* **package:** add repository ([28973b2](https://github.com/pie-framework/pie-elements/commit/28973b2))
+
+
+### Features
+
+* **dependencies:** use latest dependencies ([5220ea2](https://github.com/pie-framework/pie-elements/commit/5220ea2))
+* **function-entry:** add mathjs to function entry ([25ab268](https://github.com/pie-framework/pie-elements/commit/25ab268))
+* **function-entry:** added function entry ([5523192](https://github.com/pie-framework/pie-elements/commit/5523192))
+* **function-entry:** added hints popover ([3a6cfd3](https://github.com/pie-framework/pie-elements/commit/3a6cfd3))
+* **function-entry:** feedback and peer fixes, mathjs comparison fix ([f62ec87](https://github.com/pie-framework/pie-elements/commit/f62ec87))
+* **function-entry:** finalized for publish ([25236e3](https://github.com/pie-framework/pie-elements/commit/25236e3))
+* **function-entry:** finalized function-entry component ([6911a86](https://github.com/pie-framework/pie-elements/commit/6911a86))
+* **function-entry:** getting ready for publish, final version ([972518a](https://github.com/pie-framework/pie-elements/commit/972518a))
+* **function-entry:** hints popover fixes and adjustments ([8926f7b](https://github.com/pie-framework/pie-elements/commit/8926f7b))
+
+
+
+
+<a name="1.3.0"></a>
+# [1.3.0](https://github.com/pie-framework/pie-elements/compare/v1.2.2...v1.3.0) (2018-04-13)
+
+
+### Bug Fixes
+
+* **build:** add lint and editorconfig ([c9bce00](https://github.com/pie-framework/pie-elements/commit/c9bce00))
+
+
+### Features
+
+* **configure:** finalize configure ui ([e01b0af](https://github.com/pie-framework/pie-elements/commit/e01b0af))
+* **configure:** finalize configure ui ([37d32ff](https://github.com/pie-framework/pie-elements/commit/37d32ff))
+
+
+
+
 <a name="1.2.2"></a>
 ## 1.2.2 (2018-03-06)