+++ packages/inline-choice/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="1.5.0"></a>
+# [1.5.0](https://github.com/pie-framework/pie-elements/compare/v1.4.1...v1.5.0) (2018-04-17)
+
+
+### Bug Fixes
+
+* **dependencies:** update dependencies ([c774453](https://github.com/pie-framework/pie-elements/commit/c774453))
+* **icon:** use [@material-ui](https://github.com/material-ui)/icons package ([9e45c1a](https://github.com/pie-framework/pie-elements/commit/9e45c1a))
+
+
+
+
 <a name="1.2.2"></a>
 ## 1.2.2 (2018-03-06)