+++ packages/multiple-choice/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="1.5.0"></a>
+# [1.5.0](https://github.com/pie-framework/pie-elements/compare/v1.4.1...v1.5.0) (2018-04-17)
+
+
+### Features
+
+* **dependencies:** use latest [@pie-ui](https://github.com/pie-ui)/multiple-choice ([9b73dbb](https://github.com/pie-framework/pie-elements/commit/9b73dbb))
+
+
+
+
 <a name="1.2.2"></a>
 ## 1.2.2 (2018-03-06)